scripts/generators/docker-compose/requirements.txt
Jinja2==3.1.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
PyYAML==6.0.2
